Use job name in reported URL
We moved away from using the job name in the reported url because I
thought that job name was no longer unique. This is not quite true. If
you specify the same job multiple times it is interpretted as a variant
of the original job and so the job name is still unique.

Change-Id: I45e781530b0354c54777227cb4bffe5850848fde
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>